exec/backend/backend_python/venvproject/formychild/recommendation/serializers.py
from rest_framework import serializers
from .models import Recipe, Recommendation

class RecipeSerializer(serializers.ModelSerializer):

    class Meta:
        model = Recipe
        fields = '__all__'

class RecommendationSerializer(serializers.ModelSerializer):

    class Meta:
        model = Recommendation
        fields = '__all__'